chore: implement progressive modularity check and extract legacy tests
